+++ R/ternary_axes.R
@@ -5,6 +5,8 @@ NULL
 #' @export
 #' @rdname ternary_axis
 ternary_axis <- function(side, at = NULL, labels = TRUE, tick = TRUE,
+                         center = getOption("isopleuros.center"),
+                         scale = getOption("isopleuros.scale"),
                          font = NA, lty = "solid",
                          lwd = 1, lwd.ticks = lwd,
                          col = NULL, col.ticks = NULL, ...) {
@@ -19,25 +21,28 @@ ternary_axis <- function(side, at = NULL, labels = TRUE, tick = TRUE,
   ## Ticks and labels position
   if (is.null(at)) {
     at <- seq(from = 0, to = 1, length.out = graphics::par("xaxp")[3L] + 1)
-    at <- at[-c(1, length(at))]
+    at <- at[!(at == 0 | at == 1)]
   }
 
   axis_degree <- c(120, 240, 0)[side]
   axis_radian <- c(0, 2 * pi / 3, 4 * pi / 3)[side]
 
-  tick_start <- matrix(c(at, rep(0, length(at))), ncol = 2)
-  tick_start <- rotate(t(tick_start), theta = axis_radian)
+  pos <- matrix(data = 0, nrow = length(at), ncol = 3)
+  pos[, side] <- at
+  pos[, c(2, 3, 1)[side]] <- 1 - at
+  pos <- coordinates_ternary(pos, center = center, scale = scale)
 
   h <- abs(tcl * graphics::strheight("1", cex = 1))
   dx <- sin(pi / 6) * h
   dy <- cos(pi / 6) * h
-  tick_end <- matrix(c(at + dx, rep(-dy, length(at))), ncol = 2)
-  tick_end <- rotate(t(tick_end), theta = axis_radian)
+  tick_start <- matrix(c(pos$x, pos$y), ncol = 2)
+  tick_end <- matrix(c(pos$x + dx * c(1, 1, -2)[side],
+                       pos$y + dy * c(-1, 1, 0)[side]), ncol = 2)
 
   ## Labels
-  if (labels) {
+  if (!isFALSE(labels)) {
     if (length(labels) != length(at)) labels <- round(at * 100)
-    graphics::text(x = tick_end, label = rev(labels), srt = axis_degree,
+    graphics::text(x = tick_end, label = labels, srt = axis_degree,
                    cex = cex, col = col, font = font, adj = c(1, 0.5))
   }