+++ R/ternary_points.R
@@ -8,10 +8,12 @@ NULL
 setMethod(
   f = "ternary_points",
   signature = c(x = "numeric", y = "numeric", z = "numeric"),
-  definition = function(x, y, z, type = "p", ...) {
-    coords <- coordinates_ternary(x, y, z)
-    graphics::points(x = coords, type = type, ...)
-    invisible(data.frame(x = x, y = y, z = z))
+  definition = function(x, y, z, center = FALSE, scale = FALSE, type = "p", ...) {
+    pt <- coordinates_ternary(x, y, z, center = center, scale = scale)
+    graphics::points(x = pt, type = type, ...)
+
+    pt <- utils::modifyList(pt, list(x = x, y = y, z = z))
+    invisible(pt)
   }
 )
 
@@ -21,8 +23,11 @@ setMethod(
 setMethod(
   f = "ternary_points",
   signature = c(x = "ANY", y = "missing", z = "missing"),
-  definition = function(x, type = "p", ...) {
+  definition = function(x, center = FALSE, scale = FALSE, type = "p", ...) {
     xyz <- grDevices::xyz.coords(x)
-    methods::callGeneric(x = xyz$x, y = xyz$y, z = xyz$z, type = type, ...)
+    pt <- methods::callGeneric(x = xyz$x, y = xyz$y, z = xyz$z,
+                               center = center, scale = scale,
+                               type = type, ...)
+    invisible(pt)
   }
 )