+++ R/zzz.R
@@ -1,7 +1,9 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.isopleuros <- list(
-    isopleuros.missing = FALSE
+    isopleuros.missing = FALSE,
+    isopleuros.center = NULL,
+    isopleuros.scale = NULL
   )
   toset <- !(names(op.isopleuros) %in% names(op))
   if(any(toset)) options(op.isopleuros[toset])